requirements.txt
pytest==8.0.0
python-bcb==0.3.3
pypdf==5.5.0